requirements.txt
Django==2.0.7
django-tenants==2.0.0
django-treebeard==4.3
djangorestframework==3.8.2
djangorestframework-recursive==0.1.2
